@@ -17,11 +17,12 @@
   "devDependencies": {
     "@types/chai": "*",
     "@types/mocha": "*",
+    "@types/node": "^16.11.7",
     "@typescript-eslint/eslint-plugin": "^5.3.1",
     "@typescript-eslint/parser": "^5.3.1",
     "chai": "*",
     "mocha": "*",
-    "eslint": "^8.2.0",
+    "eslint": "^8.6.0",
     "ts-loader": "^9.2.6",
     "ts-node": "^10.4.0",
     "typedoc": "^0.22.9",
@@ -29,8 +30,5 @@
     "webpack": "^5.64.0",
     "webpack-bundle-analyzer": "^4.5.0",
     "webpack-cli": "^4.9.1"
-  },
-  "dependencies": {
-    "@types/node": "^16.11.7"
   }
 }
