@@ -4,7 +4,6 @@ exports.Equation = void 0;
 const polynom_1 = require("./polynom");
 const numeric_1 = require("../numeric");
 const coefficients_1 = require("../coefficients");
-const coefficients_2 = require("../coefficients");
 class Equation {
     _left;
     _right;
@@ -18,16 +17,26 @@ class Equation {
         this._right = new polynom_1.Polynom().zero();
         this._sign = '=';
         if (equations.length === 1) {
-            if (equations[0].isEquation === true) {
+            if (equations[0] instanceof Equation) {
                 return equations[0].clone();
             }
-            else {
+            else if (typeof equations[0] === 'string') {
                 this.parse(equations[0]);
             }
         }
         else if (equations.length === 2) {
-            this.left = (equations[0] instanceof polynom_1.Polynom) ? equations[0].clone() : new polynom_1.Polynom(equations[0]);
-            this.right = (equations[1] instanceof polynom_1.Polynom) ? equations[1].clone() : new polynom_1.Polynom(equations[1]);
+            if (equations[0] instanceof polynom_1.Polynom) {
+                this.left = equations[0].clone();
+            }
+            else if (typeof equations[0] === 'string') {
+                this.left = new polynom_1.Polynom(equations[0]);
+            }
+            if (equations[1] instanceof polynom_1.Polynom) {
+                this.right = equations[1].clone();
+            }
+            else if (typeof equations[1] === 'string') {
+                this.right = new polynom_1.Polynom(equations[1]);
+            }
         }
         else {
             return this;
@@ -43,24 +52,24 @@ class Equation {
     get solution() {
         if (this._solutions.length === 1
             &&
-                (this._solutions[0] === this._real
-                    || this._solutions[0] === this._varnothing
-                    || this._solutions[0].includes('\\left'))) {
+                (this._solutions[0].tex === this._real
+                    || this._solutions[0].tex === this._varnothing
+                    || this._solutions[0].tex.includes('\\left'))) {
             return `S = ${this._solutions[0]}`;
         }
-        return `S = \\left{ ${this._solutions.join(';')} \\right}`;
+        return `S = \\left{ ${this._solutions.map(x => x.tex).join(';')} \\right}`;
     }
     get isReal() {
         if (this._solutions === undefined) {
             this.solve();
         }
-        return this._solutions[0] === this._real;
+        return this._solutions[0].tex === this._real;
     }
     get isVarnothing() {
         if (this._solutions === undefined) {
             this.solve();
         }
-        return this._solutions[0] === this._varnothing;
+        return this._solutions[0].tex === this._varnothing;
     }
     get signAsTex() {
         if (this._sign === '>=' || this._sign === '=>' || this._sign === 'geq') {
@@ -300,19 +309,19 @@ class Equation {
     letters = () => {
         return [...new Set([...this._left.letters(), ...this._right.letters()])];
     };
-    solve = (letter) => {
+    solve = () => {
         this._solutions = [];
         this._polynom = this._left.clone().subtract(this._right);
-        switch (this._polynom.degree(letter).value) {
+        switch (this._polynom.degree().value) {
             case 0:
             case 1:
-                this._solveDegree1(letter);
+                this._solveDegree1();
                 break;
             case 2:
-                this._solveDegree2(letter);
+                this._solveDegree2();
                 break;
             default:
-                this._solveDegree3plus(letter);
+                this._solveDegree3plus();
         }
         return this;
     };
@@ -337,19 +346,31 @@ class Equation {
         }
     };
     _solveDegree1 = (letter) => {
-        const m1 = this._polynom.monomByDegree(1, letter).coefficient, m0 = this._polynom.monomByDegree(0, letter).coefficient, v = m0.clone().opposed().divide(m1).display;
+        const m1 = this._polynom.monomByDegree(1, letter).coefficient, m0 = this._polynom.monomByDegree(0, letter).coefficient, v = m0.clone().opposed().divide(m1);
         let s;
         if (this.isStrictEqual()) {
             if (m1.value === 0) {
                 if (m0.value === 0) {
-                    this._solutions = [this._real];
+                    this._solutions = [{
+                            tex: this._real,
+                            value: NaN,
+                            exact: false
+                        }];
                 }
                 else {
-                    this._solutions = [this._varnothing];
+                    this._solutions = [{
+                            tex: this._varnothing,
+                            value: NaN,
+                            exact: false
+                        }];
                 }
             }
             else {
-                this._solutions = [v];
+                this._solutions = [{
+                        tex: v.display,
+                        value: v.value,
+                        exact: v
+                    }];
             }
         }
         else {
@@ -374,7 +395,11 @@ class Equation {
                     s = `\\left\\]-\\infty;${v} \\right\\${this.isAlsoEqual() ? '\\]' : '\\['}`;
                 }
             }
-            this._solutions = [s];
+            this._solutions = [{
+                    tex: s,
+                    value: NaN,
+                    exact: false
+                }];
         }
         return this._solutions;
     };
@@ -386,87 +411,161 @@ class Equation {
             realX2 = (-b + Math.sqrt(delta)) / (2 * a);
             if (delta > 1.0e5) {
                 this._solutions = [
-                    ((-b - Math.sqrt(delta)) / (2 * a)).toFixed(5),
-                    ((-b + Math.sqrt(delta)) / (2 * a)).toFixed(5)
+                    {
+                        tex: ((-b - Math.sqrt(delta)) / (2 * a)).toFixed(5),
+                        value: realX1,
+                        exact: false
+                    },
+                    {
+                        tex: ((-b + Math.sqrt(delta)) / (2 * a)).toFixed(5),
+                        value: realX2,
+                        exact: false
+                    }
                 ];
             }
             else {
-                nthDelta = new coefficients_2.Nthroot().parse(delta).reduce();
+                nthDelta = new coefficients_1.Nthroot(delta).reduce();
                 if (nthDelta.hasRadical()) {
                     let gcd = numeric_1.Numeric.gcd(b, 2 * a, nthDelta.coefficient);
                     nthDelta.coefficient = nthDelta.coefficient / gcd;
                     if (b !== 0) {
                         if (2 * a / gcd === 1) {
                             this._solutions = [
-                                `${-b / gcd} - ${nthDelta.tex}`,
-                                `${-b / gcd} + ${nthDelta.tex}`,
+                                {
+                                    tex: `${-b / gcd} - ${nthDelta.tex}`,
+                                    value: realX1,
+                                    exact: false
+                                },
+                                {
+                                    tex: `${-b / gcd} + ${nthDelta.tex}`,
+                                    value: realX2,
+                                    exact: false
+                                },
                             ];
                         }
                         else {
                             this._solutions = [
-                                `\\dfrac{${-b / gcd} - ${nthDelta.tex} }{ ${2 * a / gcd} }`,
-                                `\\dfrac{${-b / gcd} + ${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                {
+                                    tex: `\\dfrac{${-b / gcd} - ${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                    value: realX1,
+                                    exact: false
+                                },
+                                {
+                                    tex: `\\dfrac{${-b / gcd} + ${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                    value: realX2,
+                                    exact: false
+                                },
                             ];
                         }
                     }
                     else {
                         if (2 * a / gcd === 1) {
                             this._solutions = [
-                                `- ${nthDelta.tex}`,
-                                `${nthDelta.tex}`,
+                                {
+                                    tex: `- ${nthDelta.tex}`,
+                                    value: realX1,
+                                    exact: false
+                                },
+                                {
+                                    tex: `${nthDelta.tex}`,
+                                    value: realX2,
+                                    exact: false
+                                },
                             ];
                         }
                         else {
                             this._solutions = [
-                                `\\dfrac{- ${nthDelta.tex} }{ ${2 * a / gcd} }`,
-                                `\\dfrac{${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                {
+                                    tex: `\\dfrac{- ${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                    value: realX1,
+                                    exact: false
+                                },
+                                {
+                                    tex: `\\dfrac{${nthDelta.tex} }{ ${2 * a / gcd} }`,
+                                    value: realX2,
+                                    exact: false
+                                },
                             ];
                         }
                     }
                 }
                 else {
+                    const S1 = new coefficients_1.Fraction(-b - nthDelta.coefficient, 2 * a).reduce(), S2 = new coefficients_1.Fraction(-b + nthDelta.coefficient, 2 * a).reduce();
                     this._solutions = [
-                        new coefficients_1.Fraction(-b - nthDelta.coefficient, 2 * a).reduce().dfrac,
-                        new coefficients_1.Fraction(-b + nthDelta.coefficient, 2 * a).reduce().dfrac
+                        {
+                            tex: S1.dfrac,
+                            value: realX1,
+                            exact: S1
+                        },
+                        {
+                            tex: S2.dfrac,
+                            value: realX2,
+                            exact: S2
+                        }
                     ];
                 }
             }
         }
         else if (delta === 0) {
-            this._solutions = [new coefficients_1.Fraction(-b, 2 * a).reduce().dfrac];
+            const sol = new coefficients_1.Fraction(-b, 2 * a).reduce();
+            this._solutions = [{
+                    tex: sol.dfrac,
+                    value: sol.value,
+                    exact: sol
+                }];
         }
         else {
-            this._solutions = [this._varnothing];
+            this._solutions = [{
+                    tex: this._varnothing,
+                    value: NaN,
+                    exact: false
+                }];
         }
         if (!this.isStrictEqual()) {
             if (this._solutions.length === 2) {
-                sX1 = (realX1 < realX2) ? this._solutions[0] : this._solutions[1];
-                sX2 = (realX1 < realX2) ? this._solutions[1] : this._solutions[0];
+                sX1 = (realX1 < realX2) ? this._solutions[0].tex : this._solutions[1].tex;
+                sX2 = (realX1 < realX2) ? this._solutions[1].tex : this._solutions[0].tex;
                 if ((this.isGreater() && aF.sign() === 1) || (!this.isGreater() && aF.sign() === -1)) {
-                    this._solutions = [
-                        `\\left]-\\infty ; ${sX1}\\right${this.isAlsoEqual() ? ']' : '['} \\cup \\left${this.isAlsoEqual() ? '[' : ']'}${sX2};+\\infty\\right[`
+                    this._solutions = [{
+                            tex: `\\left]-\\infty ; ${sX1}\\right${this.isAlsoEqual() ? ']' : '['} \\cup \\left${this.isAlsoEqual() ? '[' : ']'}${sX2};+\\infty\\right[`,
+                            value: NaN,
+                            exact: false
+                        }
                     ];
                 }
                 else {
-                    this._solutions = [
-                        `\\left${this.isAlsoEqual() ? '[' : ']'}${sX1} ; ${sX2}\\right${this.isAlsoEqual() ? ']' : '['}`
-                    ];
+                    this._solutions = [{
+                            tex: `\\left${this.isAlsoEqual() ? '[' : ']'}${sX1} ; ${sX2}\\right${this.isAlsoEqual() ? ']' : '['}`,
+                            value: NaN,
+                            exact: false
+                        }];
                 }
             }
-            else if (this._solutions.length === 1 && this._solutions[0] !== this._varnothing) {
+            else if (this._solutions.length === 1 && this._solutions[0].tex !== this._varnothing) {
                 if (!this.isAlsoEqual()) {
                     if ((this.isGreater() && aF.sign() === 1) || (!this.isGreater() && aF.sign() === -1)) {
-                        this._solutions = [
-                            `\\left]-\\infty ; ${this._solutions[0]}\\right[ \\cup \\left]${this._solutions[0]};+\\infty\\right[`
+                        this._solutions = [{
+                                tex: `\\left]-\\infty ; ${this._solutions[0].tex}\\right[ \\cup \\left]${this._solutions[0].tex};+\\infty\\right[`,
+                                value: NaN,
+                                exact: false
+                            }
                         ];
                     }
                     else {
-                        this._solutions = [this._varnothing];
+                        this._solutions = [{
+                                tex: this._varnothing,
+                                value: NaN,
+                                exact: false
+                            }];
                     }
                 }
                 else {
                     if ((this.isGreater() && aF.sign() === 1) || (!this.isGreater() && aF.sign() === -1)) {
-                        this._solutions = [this._real];
+                        this._solutions = [{
+                                tex: this._real,
+                                value: NaN,
+                                exact: false
+                            }];
                     }
                     else {
                     }
@@ -474,17 +573,25 @@ class Equation {
             }
             else {
                 if (this.isGreater()) {
-                    this._solutions = [aF.sign() === 1 ? this._real : this._varnothing];
+                    this._solutions = [{
+                            tex: aF.sign() === 1 ? this._real : this._varnothing,
+                            value: NaN,
+                            exact: false
+                        }];
                 }
                 else {
-                    this._solutions = [aF.sign() === -1 ? this._real : this._varnothing];
+                    this._solutions = [{
+                            tex: aF.sign() === -1 ? this._real : this._varnothing,
+                            value: NaN,
+                            exact: false
+                        }];
                 }
             }
         }
         return this._solutions;
     };
-    _solveDegree3plus = (letter) => {
-        this._solutions = [letter];
+    _solveDegree3plus = () => {
+        this._solutions = [{ tex: 'solve x - not yet handled', value: NaN, exact: false }];
         return this._solutions;
     };
 }
