@@ -1,14 +1,16 @@
 import { Point } from "./point";
+import { Fraction } from "../coefficients";
 import { Equation } from "../algebra";
 import { Line } from "./line";
 export declare class Circle {
     private _center;
-    private _radius;
     private _squareRadius;
     private _cartesian;
     private _exists;
-    constructor(...values: any);
+    constructor(...values: unknown[]);
     get center(): Point;
+    get exists(): boolean;
+    get squareRadius(): Fraction;
     get radius(): {
         tex: string;
         display: string;
@@ -17,8 +19,15 @@ export declare class Circle {
     get developed(): string;
     get display(): string;
     get cartesian(): Equation;
+    clone(): Circle;
+    private _reset;
     private parse;
-    checkCircle: (P: Equation) => boolean;
+    private _calculateCartesian;
+    private _parseCopyCircle;
+    private _parseCenterAndRadius;
+    private _parseCenterAndPointThrough;
+    private _parseEquation;
+    private _parseThroughtThreePoints;
     relativePosition: (L: Line) => number;
     lineIntersection: (L: Line) => Point[];
 }
