@@ -1,17 +1,16 @@
-import { Fraction } from "../coefficients/fraction";
+import { Fraction } from "../coefficients";
 export declare class Point {
     private _x;
     private _y;
     private _exist;
-    constructor(...values: any);
-    get isPoint(): boolean;
+    constructor(...values: unknown[]);
     get x(): Fraction;
     set x(value: Fraction);
     get y(): Fraction;
     set y(value: Fraction);
     get tex(): string;
     get display(): string;
-    parse: (...values: any) => Point;
+    parse: (...values: unknown[]) => Point;
     clone: () => Point;
     zero: () => Point;
     origin: () => Point;
