@@ -1,5 +1,6 @@
 export declare type randomCoefficientConfig = {
     negative?: boolean;
+    max?: number;
     reduced?: boolean;
     zero?: boolean;
     natural?: boolean;
@@ -7,7 +8,7 @@ export declare type randomCoefficientConfig = {
 export declare type randomMonomConfig = {
     letters?: string;
     degree?: number;
-    fraction?: boolean;
+    fraction?: boolean | randomCoefficientConfig;
     zero?: boolean;
 };
 export declare type randomPolynomConfig = randomMonomConfig & {
@@ -15,4 +16,5 @@ export declare type randomPolynomConfig = randomMonomConfig & {
     factorable?: boolean;
     allowNullMonom?: boolean;
     numberOfMonoms?: number;
+    positive?: boolean;
 };
